ocr/requirements.txt
https://github.com/simonflueckiger/tesserocr-windows_build/releases/download/tesserocr-v2.8.0-tesseract-5.5.0/tesserocr-2.8.0-cp313-cp313-win_amd64.whl
pdf2image
opencv-python
numpy
pillow
# install Poppler for Windows and add /bin to PATH: https://github.com/oschwartz10612/poppler-windows/releases/
